Reestructure code
